This has to be strict to keep it from becoming recursive. I found this useful tidbit online:
"I hadn’t considered what I was doing with this code. When comparing two objects
using the non-strict comparison operator (==) PHP compares all the properties
of the objects and if they match the objects are deemed to be equal. If they
don’t match they are not equal. In effect, we have a recursive comparison of
all the properties of each object, and all their properties, etc. until we
reach basic data types like strings and integers."



git-svn-id: file:///Users/gauthierm/svnrepo/store/trunk@14728 94490950-a86c-4493-9ccb-530f72933e47