Fix the "other" provstate option. It was currently broken.
The validation in StoreProvStateFlydown was relying on $value === null in order to skip the country validation. The CheckoutPages were setting the value null, but after the StoreProvStateFlydown::validate() had already run.

It was also broken because, by the setting the $value to null on the CheckoutPages, the flydown wasn't returning to its set state if the form was submitted and there was an error.

I've switched things around so that the flydown is aware of the 'other' option and skips validation if it's selected, and also using the 'other' value on the CheckoutPages to set null on $address->provstate.

Note: there also seems to be a problem displaying the provstate on the address which I'll fix next.



git-svn-id: file:///Users/gauthierm/svnrepo/public/store/trunk@61602 94490950-a86c-4493-9ccb-530f72933e47