save the cart at the end of handleLogin() to fix an edge case causing removed items to reappear in your cart on login.
The items would reappear if they had been removed in handleAccountCartEntries() due to no saved cart and then the page relocated without saving the cart, and seeing as most sites save the cart in the layout finalize() this is easy to do. This was never an issue on sites with saved carts because the entries were not deleted. This change will lead to the cart being saved twice when logging in and not being relocated, but that seems better than having to remember to manually save the cart before relocating in all of our code.


git-svn-id: file:///Users/gauthierm/svnrepo/public/store/trunk@69737 94490950-a86c-4493-9ccb-530f72933e47