Use one transaction on the order confirmation page
I've always found it odd that we create an account for the user on the
confirmation page even if an error occurs. This was a left over of the old
e-commerce behaviour where a user was able to checkout without an
account. If the user choose to checkout with an account we would save
their details regardless of what happened with the order. This made
sense on sites where a user could conceivably be ordering new hardware,
seeds, or trophies every few weeks. Their order for some carrot seeds might
not go through but they might come back at a later date to purchase
other things.

We no longer have any sites that have that ave sort of purchasing behaviour.
Almost all Store backed sites now sell digital products where having an
account is really only useful once you've purchased the product. I no longer
believe it makes any sense to separate the account and order saving.

Additionally one transaction will fix a bug in EM:RAP where manually
renewd accounts didn't appear to be updating in Cloud Firestore.

Before this fix the commit of the first transaction would trigger an update
for the account being saved. The new subscriptions would not yet be saved so
no new subscription info was being saved to Cloud Firestore. It appeared that
the second transaction was happening fast enough that the timestamp in the
update queue wasn't changing and that the first notification would clear that
entry from the queue once the deamon finished processing it. The second
notification would then be processed but since there would be nothing in the
queue to process no update would occur. Thus, on Cloud Firestore, no new
subscription information would appear and the account would stay out of
date.