Fix issue where account would have null id after error
The duplicate method explicitly omits copying the id. If using an
existing account this will populate the session with an account having
no id on any failure in saveOrder.

It's safe to explicitly copy the id when we duplicate the account here.
New accounts will still be set to null and existing accounts will
continue to be properly signed in.